+++ index.html
@@ -41,6 +41,10 @@
 
 
 
-    <script src="script.js"></script>
+    <script type="module" src="modules/ListItem.js" ></script>
+    <script type="module" src="modules/DateHandler.js" ></script>
+    <script type="module" src="modules/UI.js" ></script>
+    <script type="module" src="modules/Validator.js" ></script>
+    <script type="module" src="script.js"></script>
 </body>
 </html>
