+++ modules/DateHandler.js
@@ -1,0 +1,15 @@
+export default class DateHandler {
+    static getNextDay() {
+        const date = new Date();
+        const nextDay = new Date( date.setDate(date.getDate() + 1) )
+        .toLocaleDateString();
+
+        return nextDay;
+    }
+
+    static getCurrentTime() {
+        const date = new Date();
+        const currentTime = date.toLocaleTimeString();
+        return currentTime;
+    }
+}
