+++ modules/ListItem.js
@@ -1,0 +1,7 @@
+export default class ListItem {
+    constructor(task, creationDate, expirationDate) {
+        this.task = task;
+        this.creationDate = creationDate;
+        this.expirationDate = expirationDate;
+    }
+}
